move scoping panel related files to their own folder
