+++ src/DetailsView/components/details-view-overlay/scoping-panel/scoping-container.tsx
@@ -1,15 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import * as React from 'react';
-
 import * as Markup from 'assessments/markup';
 import { InspectMode } from 'background/inspect-modes';
 import { ScopingInputTypes } from 'background/scoping-input-types';
-import { SelectorInputList } from '../../common/components/selector-input-list';
-import { InspectActionMessageCreator } from '../../common/message-creators/inspect-action-message-creator';
-import { ScopingActionMessageCreator } from '../../common/message-creators/scoping-action-message-creator';
-import { FeatureFlagStoreData } from '../../common/types/store-data/feature-flag-store-data';
-import { ScopingStoreData } from '../../common/types/store-data/scoping-store-data';
+import * as React from 'react';
+import { SelectorInputList } from '../../../../common/components/selector-input-list';
+import { InspectActionMessageCreator } from '../../../../common/message-creators/inspect-action-message-creator';
+import { ScopingActionMessageCreator } from '../../../../common/message-creators/scoping-action-message-creator';
+import { FeatureFlagStoreData } from '../../../../common/types/store-data/feature-flag-store-data';
+import { ScopingStoreData } from '../../../../common/types/store-data/scoping-store-data';
 
 export interface ScopingContainerProps {
     featureFlagData: FeatureFlagStoreData;
