+++ src/DetailsView/components/details-view-overlay/scoping-panel/scoping-panel.tsx
@@ -2,13 +2,12 @@
 // Licensed under the MIT License.
 import { DefaultButton } from 'office-ui-fabric-react';
 import * as React from 'react';
-
-import { InspectActionMessageCreator } from '../../common/message-creators/inspect-action-message-creator';
-import { ScopingActionMessageCreator } from '../../common/message-creators/scoping-action-message-creator';
-import { FeatureFlagStoreData } from '../../common/types/store-data/feature-flag-store-data';
-import { ScopingStoreData } from '../../common/types/store-data/scoping-store-data';
-import { DetailsViewActionMessageCreator } from '../actions/details-view-action-message-creator';
-import { GenericPanel } from './generic-panel';
+import { InspectActionMessageCreator } from '../../../../common/message-creators/inspect-action-message-creator';
+import { ScopingActionMessageCreator } from '../../../../common/message-creators/scoping-action-message-creator';
+import { FeatureFlagStoreData } from '../../../../common/types/store-data/feature-flag-store-data';
+import { ScopingStoreData } from '../../../../common/types/store-data/scoping-store-data';
+import { DetailsViewActionMessageCreator } from '../../../actions/details-view-action-message-creator';
+import { GenericPanel } from '../../generic-panel';
 import { ScopingContainer } from './scoping-container';
 
 export type ScopingPanelDeps = {
