+++ src/tests/unit/tests/DetailsView/components/scoping-container.test.tsx
@@ -1,16 +1,18 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { ScopingInputTypes } from 'background/scoping-input-types';
 import * as Enzyme from 'enzyme';
 import * as React from 'react';
 import { Mock } from 'typemoq';
-
-import { ScopingInputTypes } from 'background/scoping-input-types';
 import { SelectorInputList } from '../../../../../common/components/selector-input-list';
 import { InspectActionMessageCreator } from '../../../../../common/message-creators/inspect-action-message-creator';
 import { ScopingActionMessageCreator } from '../../../../../common/message-creators/scoping-action-message-creator';
 import { ScopingStoreData } from '../../../../../common/types/store-data/scoping-store-data';
 import { DetailsViewActionMessageCreator } from '../../../../../DetailsView/actions/details-view-action-message-creator';
-import { ScopingContainer, ScopingContainerProps } from '../../../../../DetailsView/components/scoping-container';
+import {
+    ScopingContainer,
+    ScopingContainerProps,
+} from '../../../../../DetailsView/components/details-view-overlay/scoping-panel/scoping-container';
 
 describe('ScopingContainerTest', () => {
     const scopingSelectorsStub: ScopingStoreData = {
